+++ src/utils/logging/models/DevLogger.spec.ts
@@ -1,12 +1,14 @@
 import sinon, { SinonStub } from 'sinon';
 
-import { ProdLogger } from './ProdLogger';
-import { LoggingLevelEnum } from './types';
+import { LoggingLevelEnum } from '../types';
+import * as formatLog from '../utils/formatLog';
+import { DevLogger } from './DevLogger';
 
-describe('ProdLogger', () => {
+describe('DevLogger', () => {
 	let consoleLogStub: SinonStub;
 	let processExitStub: SinonStub;
 	let dateToISOStringStub: SinonStub;
+	let formatLogStub: SinonStub;
 
 	beforeEach(() => {
 		consoleLogStub = sinon.stub(console, 'log');
@@ -15,18 +17,48 @@ describe('ProdLogger', () => {
 
 		dateToISOStringStub = sinon.stub(Date.prototype, 'toISOString');
 		dateToISOStringStub.returns('DATE');
+
+		formatLogStub = sinon.stub(formatLog, 'formatLog');
+		formatLogStub
+			.withArgs(LoggingLevelEnum.DEBUG, '[DATE] test/DEBUG: dumb_message')
+			.returns('DEBUG: dumb_message');
+		formatLogStub
+			.withArgs(LoggingLevelEnum.INFO, '[DATE] test/INFO: dumb_message')
+			.returns('INFO: dumb_message');
+		formatLogStub
+			.withArgs(LoggingLevelEnum.WARN, '[DATE] test/WARN: dumb_message')
+			.returns('WARN: dumb_message');
+		formatLogStub
+			.withArgs(LoggingLevelEnum.ERROR, '[DATE] test/ERROR: dumb_message')
+			.returns('ERROR: dumb_message');
+		formatLogStub
+			.withArgs(LoggingLevelEnum.FATAL, '[DATE] test/FATAL: dumb_message')
+			.returns('FATAL: dumb_message');
+		formatLogStub
+			.withArgs(
+				LoggingLevelEnum.ERROR,
+				'[DATE] test/TRACE: Error[dumb_error]'
+			)
+			.returns('ERROR: dumb_error');
+		formatLogStub
+			.withArgs(
+				LoggingLevelEnum.ERROR,
+				'[DATE] test/TRACE: Error\ndumb_trace'
+			)
+			.returns('ERROR: dumb_trace');
 	});
 
 	afterEach(() => {
 		consoleLogStub.restore();
 		processExitStub.restore();
 		dateToISOStringStub.restore();
+		formatLogStub.restore();
 	});
 
 	describe('log', () => {
 		describe('with error stack', () => {
 			it('debug', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.DEBUG, 'dumb_message', {
 					name: 'Error',
@@ -35,13 +67,13 @@ describe('ProdLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
+					'DEBUG: dumb_message\nERROR: dumb_trace'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('info', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.INFO, 'dumb_message', {
 					name: 'Error',
@@ -50,13 +82,13 @@ describe('ProdLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"INFO","message":"dumb_message"}'
+					'INFO: dumb_message'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('warn', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.WARN, 'dumb_message', {
 					name: 'Error',
@@ -65,13 +97,13 @@ describe('ProdLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
+					'WARN: dumb_message\nERROR: dumb_trace'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('error', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.ERROR, 'dumb_message', {
 					name: 'Error',
@@ -80,13 +112,13 @@ describe('ProdLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
+					'ERROR: dumb_message\nERROR: dumb_trace'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('fatal', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.FATAL, 'dumb_message', {
 					name: 'Error',
@@ -95,7 +127,7 @@ describe('ProdLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
+					'FATAL: dumb_message\nERROR: dumb_trace'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 
@@ -106,7 +138,7 @@ describe('ProdLogger', () => {
 
 		describe('with error message only', () => {
 			it('debug', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.DEBUG, 'dumb_message', {
 					name: 'Error',
@@ -114,13 +146,13 @@ describe('ProdLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
+					'DEBUG: dumb_message\nERROR: dumb_error'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('info', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.INFO, 'dumb_message', {
 					name: 'Error',
@@ -128,13 +160,13 @@ describe('ProdLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"INFO","message":"dumb_message"}'
+					'INFO: dumb_message'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('warn', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.WARN, 'dumb_message', {
 					name: 'Error',
@@ -142,13 +174,13 @@ describe('ProdLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
+					'WARN: dumb_message\nERROR: dumb_error'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('error', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.ERROR, 'dumb_message', {
 					name: 'Error',
@@ -156,13 +188,13 @@ describe('ProdLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
+					'ERROR: dumb_message\nERROR: dumb_error'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('fatal', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.FATAL, 'dumb_message', {
 					name: 'Error',
@@ -170,7 +202,7 @@ describe('ProdLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
+					'FATAL: dumb_message\nERROR: dumb_error'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 
@@ -181,56 +213,56 @@ describe('ProdLogger', () => {
 
 		describe('without error', () => {
 			it('debug', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.DEBUG, 'dumb_message');
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message"}'
+					'DEBUG: dumb_message'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('info', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.INFO, 'dumb_message');
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"INFO","message":"dumb_message"}'
+					'INFO: dumb_message'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('warn', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.WARN, 'dumb_message');
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message"}'
+					'WARN: dumb_message'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('error', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.ERROR, 'dumb_message');
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message"}'
+					'ERROR: dumb_message'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('fatal', function () {
-				const devLogger = new ProdLogger('test');
+				const devLogger = new DevLogger('test');
 
 				devLogger.log(LoggingLevelEnum.FATAL, 'dumb_message');
 
 				consoleLogStub.should.have.been.calledWith(
-					'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message"}'
+					'FATAL: dumb_message'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 
@@ -242,7 +274,7 @@ describe('ProdLogger', () => {
 
 	describe('debug', () => {
 		it('with error stack', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.debug('dumb_message', {
 				name: 'Error',
@@ -251,13 +283,13 @@ describe('ProdLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
+				'DEBUG: dumb_message\nERROR: dumb_trace'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('with error message only', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.debug('dumb_message', {
 				name: 'Error',
@@ -265,37 +297,33 @@ describe('ProdLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
+				'DEBUG: dumb_message\nERROR: dumb_error'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('without error', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.debug('dumb_message');
 
-			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message"}'
-			);
+			consoleLogStub.should.have.been.calledWith('DEBUG: dumb_message');
 			consoleLogStub.should.have.been.calledOnce();
 		});
 	});
 
 	it('info', function () {
-		const devLogger = new ProdLogger('test');
+		const devLogger = new DevLogger('test');
 
 		devLogger.info('dumb_message');
 
-		consoleLogStub.should.have.been.calledWith(
-			'{"date":"DATE","origin":"test","type":"INFO","message":"dumb_message"}'
-		);
+		consoleLogStub.should.have.been.calledWith('INFO: dumb_message');
 		consoleLogStub.should.have.been.calledOnce();
 	});
 
 	describe('warn', () => {
 		it('with error stack', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.warn('dumb_message', {
 				name: 'Error',
@@ -304,13 +332,13 @@ describe('ProdLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
+				'WARN: dumb_message\nERROR: dumb_trace'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('with error message only', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.warn('dumb_message', {
 				name: 'Error',
@@ -318,26 +346,24 @@ describe('ProdLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
+				'WARN: dumb_message\nERROR: dumb_error'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('without error', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.warn('dumb_message');
 
-			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message"}'
-			);
+			consoleLogStub.should.have.been.calledWith('WARN: dumb_message');
 			consoleLogStub.should.have.been.calledOnce();
 		});
 	});
 
 	describe('error', () => {
 		it('with error stack', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.error('dumb_message', {
 				name: 'Error',
@@ -346,13 +372,13 @@ describe('ProdLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
+				'ERROR: dumb_message\nERROR: dumb_trace'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('with error message only', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.error('dumb_message', {
 				name: 'Error',
@@ -360,26 +386,24 @@ describe('ProdLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
+				'ERROR: dumb_message\nERROR: dumb_error'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('without error', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.error('dumb_message');
 
-			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message"}'
-			);
+			consoleLogStub.should.have.been.calledWith('ERROR: dumb_message');
 			consoleLogStub.should.have.been.calledOnce();
 		});
 	});
 
 	describe('fatal', () => {
 		it('with error stack', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.fatal('dumb_message', {
 				name: 'Error',
@@ -388,7 +412,7 @@ describe('ProdLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
+				'FATAL: dumb_message\nERROR: dumb_trace'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 
@@ -397,7 +421,7 @@ describe('ProdLogger', () => {
 		});
 
 		it('with error message only', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.fatal('dumb_message', {
 				name: 'Error',
@@ -405,7 +429,7 @@ describe('ProdLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
+				'FATAL: dumb_message\nERROR: dumb_error'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 
@@ -414,13 +438,11 @@ describe('ProdLogger', () => {
 		});
 
 		it('without error', function () {
-			const devLogger = new ProdLogger('test');
+			const devLogger = new DevLogger('test');
 
 			devLogger.fatal('dumb_message');
 
-			consoleLogStub.should.have.been.calledWith(
-				'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message"}'
-			);
+			consoleLogStub.should.have.been.calledWith('FATAL: dumb_message');
 			consoleLogStub.should.have.been.calledOnce();
 
 			processExitStub.should.have.been.calledWith(1);
