+++ src/utils/logging/models/ProdLogger.spec.ts
@@ -1,14 +1,12 @@
 import sinon, { SinonStub } from 'sinon';
 
-import { DevLogger } from './DevLogger';
-import * as formatLog from './formatLog';
-import { LoggingLevelEnum } from './types';
+import { LoggingLevelEnum } from '../types';
+import { ProdLogger } from './ProdLogger';
 
-describe('DevLogger', () => {
+describe('ProdLogger', () => {
 	let consoleLogStub: SinonStub;
 	let processExitStub: SinonStub;
 	let dateToISOStringStub: SinonStub;
-	let formatLogStub: SinonStub;
 
 	beforeEach(() => {
 		consoleLogStub = sinon.stub(console, 'log');
@@ -17,48 +15,18 @@ describe('DevLogger', () => {
 
 		dateToISOStringStub = sinon.stub(Date.prototype, 'toISOString');
 		dateToISOStringStub.returns('DATE');
-
-		formatLogStub = sinon.stub(formatLog, 'formatLog');
-		formatLogStub
-			.withArgs(LoggingLevelEnum.DEBUG, '[DATE] test/DEBUG: dumb_message')
-			.returns('DEBUG: dumb_message');
-		formatLogStub
-			.withArgs(LoggingLevelEnum.INFO, '[DATE] test/INFO: dumb_message')
-			.returns('INFO: dumb_message');
-		formatLogStub
-			.withArgs(LoggingLevelEnum.WARN, '[DATE] test/WARN: dumb_message')
-			.returns('WARN: dumb_message');
-		formatLogStub
-			.withArgs(LoggingLevelEnum.ERROR, '[DATE] test/ERROR: dumb_message')
-			.returns('ERROR: dumb_message');
-		formatLogStub
-			.withArgs(LoggingLevelEnum.FATAL, '[DATE] test/FATAL: dumb_message')
-			.returns('FATAL: dumb_message');
-		formatLogStub
-			.withArgs(
-				LoggingLevelEnum.ERROR,
-				'[DATE] test/TRACE: Error[dumb_error]'
-			)
-			.returns('ERROR: dumb_error');
-		formatLogStub
-			.withArgs(
-				LoggingLevelEnum.ERROR,
-				'[DATE] test/TRACE: Error\ndumb_trace'
-			)
-			.returns('ERROR: dumb_trace');
 	});
 
 	afterEach(() => {
 		consoleLogStub.restore();
 		processExitStub.restore();
 		dateToISOStringStub.restore();
-		formatLogStub.restore();
 	});
 
 	describe('log', () => {
 		describe('with error stack', () => {
 			it('debug', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.DEBUG, 'dumb_message', {
 					name: 'Error',
@@ -67,13 +35,13 @@ describe('DevLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'DEBUG: dumb_message\nERROR: dumb_trace'
+					'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('info', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.INFO, 'dumb_message', {
 					name: 'Error',
@@ -82,13 +50,13 @@ describe('DevLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'INFO: dumb_message'
+					'{"date":"DATE","origin":"test","type":"INFO","message":"dumb_message"}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('warn', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.WARN, 'dumb_message', {
 					name: 'Error',
@@ -97,13 +65,13 @@ describe('DevLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'WARN: dumb_message\nERROR: dumb_trace'
+					'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('error', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.ERROR, 'dumb_message', {
 					name: 'Error',
@@ -112,13 +80,13 @@ describe('DevLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'ERROR: dumb_message\nERROR: dumb_trace'
+					'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('fatal', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.FATAL, 'dumb_message', {
 					name: 'Error',
@@ -127,7 +95,7 @@ describe('DevLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'FATAL: dumb_message\nERROR: dumb_trace'
+					'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 
@@ -138,7 +106,7 @@ describe('DevLogger', () => {
 
 		describe('with error message only', () => {
 			it('debug', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.DEBUG, 'dumb_message', {
 					name: 'Error',
@@ -146,13 +114,13 @@ describe('DevLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'DEBUG: dumb_message\nERROR: dumb_error'
+					'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('info', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.INFO, 'dumb_message', {
 					name: 'Error',
@@ -160,13 +128,13 @@ describe('DevLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'INFO: dumb_message'
+					'{"date":"DATE","origin":"test","type":"INFO","message":"dumb_message"}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('warn', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.WARN, 'dumb_message', {
 					name: 'Error',
@@ -174,13 +142,13 @@ describe('DevLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'WARN: dumb_message\nERROR: dumb_error'
+					'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('error', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.ERROR, 'dumb_message', {
 					name: 'Error',
@@ -188,13 +156,13 @@ describe('DevLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'ERROR: dumb_message\nERROR: dumb_error'
+					'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('fatal', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.FATAL, 'dumb_message', {
 					name: 'Error',
@@ -202,7 +170,7 @@ describe('DevLogger', () => {
 				});
 
 				consoleLogStub.should.have.been.calledWith(
-					'FATAL: dumb_message\nERROR: dumb_error'
+					'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 
@@ -213,56 +181,56 @@ describe('DevLogger', () => {
 
 		describe('without error', () => {
 			it('debug', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.DEBUG, 'dumb_message');
 
 				consoleLogStub.should.have.been.calledWith(
-					'DEBUG: dumb_message'
+					'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message"}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('info', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.INFO, 'dumb_message');
 
 				consoleLogStub.should.have.been.calledWith(
-					'INFO: dumb_message'
+					'{"date":"DATE","origin":"test","type":"INFO","message":"dumb_message"}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('warn', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.WARN, 'dumb_message');
 
 				consoleLogStub.should.have.been.calledWith(
-					'WARN: dumb_message'
+					'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message"}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('error', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.ERROR, 'dumb_message');
 
 				consoleLogStub.should.have.been.calledWith(
-					'ERROR: dumb_message'
+					'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message"}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 			});
 
 			it('fatal', function () {
-				const devLogger = new DevLogger('test');
+				const devLogger = new ProdLogger('test');
 
 				devLogger.log(LoggingLevelEnum.FATAL, 'dumb_message');
 
 				consoleLogStub.should.have.been.calledWith(
-					'FATAL: dumb_message'
+					'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message"}'
 				);
 				consoleLogStub.should.have.been.calledOnce();
 
@@ -274,7 +242,7 @@ describe('DevLogger', () => {
 
 	describe('debug', () => {
 		it('with error stack', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.debug('dumb_message', {
 				name: 'Error',
@@ -283,13 +251,13 @@ describe('DevLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'DEBUG: dumb_message\nERROR: dumb_trace'
+				'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('with error message only', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.debug('dumb_message', {
 				name: 'Error',
@@ -297,33 +265,37 @@ describe('DevLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'DEBUG: dumb_message\nERROR: dumb_error'
+				'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('without error', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.debug('dumb_message');
 
-			consoleLogStub.should.have.been.calledWith('DEBUG: dumb_message');
+			consoleLogStub.should.have.been.calledWith(
+				'{"date":"DATE","origin":"test","type":"DEBUG","message":"dumb_message"}'
+			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 	});
 
 	it('info', function () {
-		const devLogger = new DevLogger('test');
+		const devLogger = new ProdLogger('test');
 
 		devLogger.info('dumb_message');
 
-		consoleLogStub.should.have.been.calledWith('INFO: dumb_message');
+		consoleLogStub.should.have.been.calledWith(
+			'{"date":"DATE","origin":"test","type":"INFO","message":"dumb_message"}'
+		);
 		consoleLogStub.should.have.been.calledOnce();
 	});
 
 	describe('warn', () => {
 		it('with error stack', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.warn('dumb_message', {
 				name: 'Error',
@@ -332,13 +304,13 @@ describe('DevLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'WARN: dumb_message\nERROR: dumb_trace'
+				'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('with error message only', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.warn('dumb_message', {
 				name: 'Error',
@@ -346,24 +318,26 @@ describe('DevLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'WARN: dumb_message\nERROR: dumb_error'
+				'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('without error', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.warn('dumb_message');
 
-			consoleLogStub.should.have.been.calledWith('WARN: dumb_message');
+			consoleLogStub.should.have.been.calledWith(
+				'{"date":"DATE","origin":"test","type":"WARN","message":"dumb_message"}'
+			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 	});
 
 	describe('error', () => {
 		it('with error stack', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.error('dumb_message', {
 				name: 'Error',
@@ -372,13 +346,13 @@ describe('DevLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'ERROR: dumb_message\nERROR: dumb_trace'
+				'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('with error message only', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.error('dumb_message', {
 				name: 'Error',
@@ -386,24 +360,26 @@ describe('DevLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'ERROR: dumb_message\nERROR: dumb_error'
+				'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 
 		it('without error', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.error('dumb_message');
 
-			consoleLogStub.should.have.been.calledWith('ERROR: dumb_message');
+			consoleLogStub.should.have.been.calledWith(
+				'{"date":"DATE","origin":"test","type":"ERROR","message":"dumb_message"}'
+			);
 			consoleLogStub.should.have.been.calledOnce();
 		});
 	});
 
 	describe('fatal', () => {
 		it('with error stack', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.fatal('dumb_message', {
 				name: 'Error',
@@ -412,7 +388,7 @@ describe('DevLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'FATAL: dumb_message\nERROR: dumb_trace'
+				'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message","error":{"name":"Error","message":"dumb_error","trace":"dumb_trace"}}'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 
@@ -421,7 +397,7 @@ describe('DevLogger', () => {
 		});
 
 		it('with error message only', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.fatal('dumb_message', {
 				name: 'Error',
@@ -429,7 +405,7 @@ describe('DevLogger', () => {
 			});
 
 			consoleLogStub.should.have.been.calledWith(
-				'FATAL: dumb_message\nERROR: dumb_error'
+				'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message","error":{"name":"Error","message":"dumb_error"}}'
 			);
 			consoleLogStub.should.have.been.calledOnce();
 
@@ -438,11 +414,13 @@ describe('DevLogger', () => {
 		});
 
 		it('without error', function () {
-			const devLogger = new DevLogger('test');
+			const devLogger = new ProdLogger('test');
 
 			devLogger.fatal('dumb_message');
 
-			consoleLogStub.should.have.been.calledWith('FATAL: dumb_message');
+			consoleLogStub.should.have.been.calledWith(
+				'{"date":"DATE","origin":"test","type":"FATAL","message":"dumb_message"}'
+			);
 			consoleLogStub.should.have.been.calledOnce();
 
 			processExitStub.should.have.been.calledWith(1);
